处理tsf光标跟随问题。

+++ sinstar3_tsf/keyeventsink.cpp
@@ -16,7 +16,7 @@ STDAPI CSinstar3Tsf::OnTestKeyDown(ITfContext *pContext, WPARAM wParam, LPARAM l
 		return S_OK;
 	}
 
-	m_pSinstar3->ProcessKeyStoke(pContext,wParam,lParam, TRUE, pfEaten);
+	m_pSinstar3->ProcessKeyStoke(pContext,(UINT)wParam,lParam, TRUE, pfEaten);
 
 	SLOGFMTF("OnTestKeyDown: wParam:%08x,lparam:%08x",wParam,lParam);
     return S_OK;
@@ -30,7 +30,7 @@ STDAPI CSinstar3Tsf::OnTestKeyUp(ITfContext *pContext, WPARAM wParam, LPARAM lPa
 		return S_OK;
 	}
 
-	m_pSinstar3->ProcessKeyStoke(pContext,wParam, lParam,FALSE, pfEaten);
+	m_pSinstar3->ProcessKeyStoke(pContext,(UINT)wParam, lParam,FALSE, pfEaten);
 
 	return S_OK;
 }
@@ -45,7 +45,7 @@ STDAPI CSinstar3Tsf::OnKeyDown(ITfContext *pContext, WPARAM wParam, LPARAM lPara
 	OnTestKeyDown(pContext,wParam,lParam,pfEaten);
  	if(*pfEaten)
 	{
-		m_pSinstar3->TranslateKey(pContext,wParam, MapVirtualKey(wParam,0), TRUE, pfEaten);
+		m_pSinstar3->TranslateKey(pContext,(UINT)wParam, MapVirtualKey((UINT)wParam,0), TRUE, pfEaten);
 	}
 	_bInKeyProc=FALSE;
 	return S_OK;
@@ -58,7 +58,7 @@ STDAPI CSinstar3Tsf::OnKeyUp(ITfContext *pContext, WPARAM wParam, LPARAM lParam,
 	OnTestKeyUp(pContext,wParam,lParam,pfEaten);
 	if(*pfEaten)
 	{
-		m_pSinstar3->TranslateKey(pContext,wParam, MapVirtualKey(wParam,0), FALSE, pfEaten);
+		m_pSinstar3->TranslateKey(pContext,(UINT)wParam, MapVirtualKey((UINT)wParam,0), FALSE, pfEaten);
 	}
 	_bInKeyProc=FALSE;
